pyproject.toml
[project]
name = "applybot"
version = "1.0.0"
description = "Automated job application system for hh.ru with AI-powered cover letters"
authors = [
    {name = "Ivan Tkachenko", email = "i.tkachenko@zohomail.eu"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11,<3.12"
keywords = ["hh.ru", "job-application", "automation", "ai", "fastapi"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
]
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn>=0.34.0,<1.0.0",
    "sqlalchemy>=2.0.0,<3.0.0",
    "anthropic>=0.54.0,<1.0.0",
    "pydantic-settings>=2.9.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "psycopg2-binary>=2.9.0,<3.0.0",
    "asyncpg>=0.30.0,<1.0.0",
    "alembic>=1.16.0,<2.0.0",
    "apscheduler>=3.10.0,<4.0.0",
    "httpx>=0.28.0,<1.0.0",
]

[project.urls]
Repository = "https://github.com/FrostWillmott/ApplyBot"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{ include = "app", from = "." }]

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.0"
mypy = "^1.16.0"
pre-commit = "^4.2.0"
pytest = "^8.3.0"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.24.0"
aiosqlite = "^0.20.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-v --tb=short --cov=app --cov-report=term-missing"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["app"]
omit = [
    "app/tasks.py",
    "app/services/hh_client.py",
    "app/services/scheduler_service.py",
    "app/services/llm/enhanced_providers.py",
    "app/services/llm/providers.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false
ignore_missing_imports = true
